requirements.txt
aliyun-python-sdk-core==2.13.33
aliyun-python-sdk-kms==2.14.0
asgiref==3.3.1
astor==0.8.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crcmod==1.7
cryptography==3.4.6
cycler==0.10.0
decorator==5.0.5
Django==3.1.7
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-oss-storage==1.1.1
gast==0.3.3
idna==2.10
jmespath==0.10.0
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib @ file:///D:/Java/matplotlib-3.3.4-cp38-cp38-win_amd64.whl
numpy==1.20.2
oss2==2.14.0
paddlepaddle==2.0.0
Pillow==8.1.2
protobuf==3.15.7
pycparser==2.20
pycryptodome==3.10.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
